Continued integration of core data
Have effectively got back to where I was before adding core data, in
that all my tests now work again.  Learnt a valuable lesson about not
adding target membership!

+++ Simple Accounts/TransactionCategory+CoreDataProperties.swift
@@ -14,7 +14,7 @@ import CoreData
 
 extension TransactionCategory {
 
-    typealias IconIndex = Int
+    typealias IconIndex = Int16
     
     @NSManaged var name: String
     @NSManaged var icon: IconIndex
